third_party/blink/web_tests/platform/generic/fast/js/mozilla/strict/8.12.7-expected.txt

PASS 'use strict'; var o = setup(); delete o.all is true
PASS var o = setup(); delete o.all is true
PASS true === true
PASS 'use strict'; var o = setup(); delete o.nowrite is true
PASS var o = setup(); delete o.nowrite is true
PASS true === true
PASS 'use strict'; var o = setup(); delete o.noconfig threw exception of type TypeError.
PASS var o = setup(); delete o.noconfig is false
PASS true === true
PASS 'use strict'; var o = setup(); delete o.noble threw exception of type TypeError.
PASS var o = setup(); delete o.noble is false
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
